added building flag

+++ server/plugins/tokens.ts
@@ -1,6 +1,9 @@
 import { prisma } from "../utils/prisma";
 
 export default defineNitroPlugin((nitro) => {
+    const runtimeConfig = useRuntimeConfig();
+    if (runtimeConfig.build) return;
+
     setInterval(async () => {
         await prisma.apiToken.deleteMany({
             where: {
